+++ src/acetime/zonedb/zone_policies.py
@@ -28739,4 +28739,3 @@ ZONE_POLICY_Zion: ZonePolicy = {
 #   SAVE '2:00' different from 1:00,
 # }
 
-